llm_requirements.txt
accelerate==0.20.3
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
albumentations==1.3.1
altair==5.0.1
antlr4-python3-runtime==4.9.3
anyio==3.7.0
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1670263926556/work
async-timeout==4.0.2
attrs==23.1.0
backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
backports.functools-lru-cache @ file:///home/conda/feedstock_root/build_artifacts/backports.functools_lru_cache_1687772187254/work
backports.zoneinfo==0.2.1
blinker==1.6.2
blis==0.7.9
braceexpand==0.1.7
cachetools==5.3.1
catalogue==2.0.8
category-encoders==2.6.1
certifi==2023.5.7
cfgv==3.3.1
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
cmake==3.26.4
confection==0.0.4
contexttimer==0.3.3
contourpy==1.1.0
cycler==0.11.0
cymem==2.0.7
dask==2023.5.0
dataclasses==0.6
debugpy @ file:///tmp/build/80754af9/debugpy_1637091796427/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
decord==0.6.0
distlib==0.3.6
docker-pycreds==0.4.0
einops==0.6.1
entrypoints @ file:///home/conda/feedstock_root/build_artifacts/entrypoints_1643888246732/work
exceptiongroup==1.1.1
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1667317341051/work
ExifRead-nocycle==3.0.1
fairscale==0.4.4
fastapi==0.97.0
ffmpy==0.3.0
filelock==3.12.0
fire==0.4.0
fonttools==4.40.0
frozenlist==1.3.3
fschat @ git+https://github.com/lm-sys/FastChat.git@f34f28cedcb8906fd026f22ec3ef41435a8e24ac
fsspec==2022.11.0
ftfy==6.1.1
gitdb==4.0.10
GitPython==3.1.31
gradio==3.23.0
gradio_client==0.2.7
h11==0.14.0
hiq-python==1.1.12
httpcore==0.17.2
httpx==0.24.1
huggingface-hub==0.15.1
identify==2.5.24
idna==3.4
imageio==2.31.1
img2dataset==1.41.0
importlib-metadata==6.7.0
importlib-resources==5.12.0
iopath==0.1.10
ipykernel @ file:///home/conda/feedstock_root/build_artifacts/ipykernel_1655369107642/work
ipython==8.12.2
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1669134318875/work
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
jupyter-client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1633454794268/work
jupyter_core @ file:///home/conda/feedstock_root/build_artifacts/jupyter_core_1669775081186/work
kaggle==1.5.13
kaleido==0.2.1
kiwisolver==1.4.4
langcodes==3.3.0
lazy_loader==0.2
linkify-it-py==2.0.2
lit==16.0.5.post0
-e git+ssh://git@github.com/facebookresearch/llama.git@57b0eb62de0636e75af471e49e2f1862d908d9d8#egg=llama
locket==1.0.0
markdown-it-py==2.2.0
markdown2==2.4.9
MarkupSafe==2.1.3
matplotlib==3.7.1
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1660814786464/work
mdit-py-plugins==0.3.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
murmurhash==1.0.9
nest-asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1664684991461/work
networkx==3.1
nodeenv==1.8.0
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
omegaconf==2.3.0
opencv-python-headless==4.5.5.64
opendatasets==0.1.22
orjson==3.9.1
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1681337016113/work
pandas==1.5.3
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1638334955874/work
partd==1.4.0
pathtools==0.1.2
pathy==0.10.2
patsy==0.5.3
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1667297516076/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
platformdirs==3.7.0
plotly==5.15.0
portalocker==2.7.0
pre-commit==3.3.3
preshed==3.0.8
promise==2.3
prompt-toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1677600924538/work
protobuf==3.20.3
psutil==5.9.5
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1609419310487/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1642875951954/work
py-itree==0.0.19
pyarrow==7.0.0
pycocoevalcap==1.2
pycocotools==2.0.6
pydantic==1.10.9
pydeck==0.8.1b0
pydub==0.25.1
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1681904169130/work
pyllama==0.0.9
Pympler==1.0.1
pyparsing==3.1.0
pyrsistent==0.19.3
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
python-magic==0.4.27
python-multipart==0.0.6
python-slugify==8.0.1
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==19.0.2
qudida==0.0.4
regex==2023.6.3
requests==2.31.0
rich==13.4.1
safetensors==0.3.1
-e git+ssh://git@github.com/VPGTrans/VPGTrans.git@a685f912bb9ebc4079b6eaf31b435927590f45fe#egg=salesforce_lavis
scikit-image==0.21.0
scikit-learn==1.2.2
scipy==1.10.1
semantic-version==2.10.0
sentencepiece==0.1.97
sentry-sdk==1.25.1
setproctitle==1.3.2
shortuuid==1.0.11
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
smart-open==6.3.0
smmap==5.0.0
sniffio==1.3.0
spacy==3.5.3
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.6
stack-data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1669632077133/work
starlette==0.27.0
statsmodels==0.14.0
streamlit==1.23.1
svgwrite==1.4.3
sympy==1.12
tenacity==8.2.2
termcolor==2.3.0
text-unidecode==1.3
thinc==8.1.10
threadpoolctl==3.1.0
tifffile==2023.4.12
timm==0.4.12
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
torch==2.0.1
torchvision==0.15.2
tornado==6.3.2
tqdm==4.65.0
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1675110562325/work
transformers @ git+https://github.com/huggingface/transformers.git@abaca9f9432a84cfaa95531de4c72334f38a42f2
triton==2.0.0
typer==0.7.0
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1685704949284/work
tzdata==2023.3
tzlocal==4.3
uc-micro-py==1.0.2
urllib3==2.0.3
uvicorn==0.22.0
validators==0.20.0
virtualenv==20.23.1
wandb==0.12.21
wasabi==1.1.2
watchdog==3.0.0
wavedrom==2.0.3.post3
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1673864653149/work
webdataset==0.2.48
websockets==11.0.3
yarl==1.9.2
zipp==3.15.0
